chore(tests): reorganize into utils/editor/perf
