+++ package.json
@@ -41,7 +41,7 @@
     "test-browser-collab-watch": "npm run build && FORCE_WEBSOCKET=true VITE_ENABLE_FAST_REFRESH=false node ./scripts/playwright-watch.mjs --project=chromium",
     "test-browser-watch": "npm run build && VITE_ENABLE_FAST_REFRESH=false node ./scripts/playwright-watch.mjs --project=chromium",
     "test-unit": "npx vitest run",
-    "test-performance": "VITE_PERFORMANCE_TESTS=true VITE_LOG_LEVEL=info time -f 'Time: %es' npx vitest run performance/performance --reporter ./tests/unit/performance/reporter.ts -t",
+    "test-performance": "VITE_PERFORMANCE_TESTS=true VITE_LOG_LEVEL=info time -f 'Time: %es' npx vitest run perf/performance --reporter ./tests/unit/perf/reporter.ts -t",
     "test-unit-collab": "FORCE_WEBSOCKET=true npx vitest run tests/unit/collab --maxWorkers=1",
     "test-unit-watch": "npx vitest --watch --ui",
     "test-unit-minimal": "npx vitest --reporter=./tests/unit/common/reporter.ts --watch --ui",
