+++ tests/unit/editor/collab/provider.spec.ts
@@ -2,7 +2,7 @@ import { afterAll, beforeAll, expect, it } from "vitest";
 import { WebsocketProvider } from "y-websocket";
 import * as Y from "yjs";
 import WS from "ws";
-import { env } from "../../../config/env.server";
+import { env } from "../../../../config/env.server";
 
 declare global {
   // eslint-disable-next-line no-var
