+++ tests/unit/editor/regression/broken_schema.spec.ts
@@ -1,4 +1,4 @@
-import "../common";
+import "../../common";
 import { Note } from "@/components/Editor/plugins/remdo/utils/api";
 import { $isListItemNode } from "@lexical/list";
 import { $getEditor, $isTextNode } from "lexical";
